% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoDiatomRiverData.R
\docType{data}
\name{demoDiatomRiverData}
\alias{demoDiatomRiverData}
\title{Demo data for river diatom sites}
\description{
Demo data for river diatom sites
}
\examples{
\dontrun{
demoDiatomRiverData <- demoDiatomRiverData
}
}
\references{
\url{www.sepa.org.uk}
}
\keyword{datasets}
