% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytobenthosMetrics.R
\docType{data}
\name{LookUpDARLEQ2}
\alias{LookUpDARLEQ2}
\title{DARLEQ. TDIs associated with the presence of diatoms.}
\format{A data frame with 1275 rows and 24 variables}
\description{
TDI (Trophic Diatom Index) associated with diatoms in rivers and lakes,
providing a information on the nutrient status of rivers, as appear in DARLEQ2
}
\details{
The columns are as follows:

\itemize{
 \item TaxonId - Code added by tool developer
 \item NBSCode - NBN Atlas idenfier code  - not necessarily the preferred NBN Atlas code
 \item AccCode - ?
 \item Class
 \item Family
 \item Authority
 \item Genus
 \item TaxonName
 \item DARESTaxonName - Name used in DARES tool?
 \item DARESAuthority - Authority used in DARES tool?
 \item Aggregate - ?
 \item Status - ?
 \item TDIo - First diatom metric?
 \item TDI3 - DARES river metric?
 \item TDI4 - DARLEQ river metric
 \item LTDI1 - Previous Lake metric
 \item LTDI3 - DARLEQ lake metric
 \item DAM - Diatom Acid Metric
 \item Motile - Taxa classed as motile
 \item OrganicTolerant - Taxa classed as organic tolerant
 \item Planktic - Taxa classed as planktic
 \item Saline- Taxa classed as saline tolerant
 \item Notes - For example if typo or name change
}
}
\keyword{datasets}
