% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytobenthosMetrics.R
\docType{package}
\name{phytobenthosMetrics}
\alias{phytobenthosMetrics}
\alias{phytobenthosMetrics-package}
\title{Phytobenthos Biotic indices}
\description{
Metrics for assessing ecological indicators of biodiversity in rivers and
lakes.
Specifically coded to be used with TIBCO Enterprise Runtime for R (TERR).
}
\details{
\tabular{ll}{
 Package: \tab phytobenthosMetrics\cr
 Type: \tab Package\cr
 Version: \tab 0.98-2\cr
 Date: \tab 2017-03-30\cr
 License: \tab MIT\cr
}
This package contains several functions which compute multiple ecological
metrics using multiple season combinations.
\itemize{
 \item DARLEQ (\code{\link{CalcDARLEQ}})\cr Diatoms for Assessing River and
 Lake Ecological Quality
}
}
\author{
Scottish Environment Protection Agency
Maintainer: Carlos Ruiz \email{carlos.ruiz@sepa.org.uk}
}
