% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoDiatomRiverData.R
\docType{data}
\name{demoDiatomRiverData}
\alias{demoDiatomRiverData}
\title{Demo data for river diatom sites}
\format{A data frame with 31891 rows and 8 variables
\describe{
  \item{Type}{R for River}
  \item{SampleID}{Sample identifier}
  \item{SiteID}{Site identifier}
  \item{Date}{Date}
  \item{Taxon.Code}{Taxon identifier}
  \item{Taxon.name}{Taxon name}
  \item{Abundance}{Count of abundance on diatom slide}
  \item{Alkalinity}{Average alkalinity of site}
}}
\usage{
demoDiatomRiverData
}
\description{
A dataset containing 1257 dummy diatom samples based on data from Scottish
rivers diamonds.
}
\examples{
\dontrun{
demoDiatomRiverData <- demoDiatomRiverData
}
}
\references{
\url{www.sepa.org.uk}
}
\keyword{datasets}
