% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytobenthosMetrics.R
\docType{data}
\name{LookUpDARLEQ2}
\alias{LookUpDARLEQ2}
\title{DARLEQ. TDIs associated with the presence of diatoms.
TDI (Trophic Diatom Index) associated with diatoms in rivers and lakes,
providing a information on the nutrient status of rivers, as appear in DARLEQ2}
\format{A data frame with 1275 rows and 24 variables}
\description{
The columns are as follows:
}
\details{
\itemize{
 \item TaxonId - Code added by tool developer. This column should have unique codes
 and all rows in the column should have an entry.
 \item NBSCode - \href{https://api.nbnatlas.org/;jsessionid=821F3DBADB0ABD14F10EA7B2354EE741#ws80}{NBN Atlas identifier code}
 more specifically the \href{http://nbn-sd-dev.nhm.ac.uk/glossary.php}{'TVK'}
 code created by the Natural History Museum. This is not necessarily the
 preferred NBN Atlas code. These are GUIDs, each taxon concept should have a
 unique GUID/code, therefore each row in this table has a unique
 NBSCode. However, not all taxa have in the table have been added to the
 NHM system. Therefore this column is either Null or has a unique value.
 \item AccCode
 \item Class
 \item Family
 \item Authority
 \item Genus
 \item TaxonName
 \item DARESTaxonName - Name used in DARES tool?
 \item DARESAuthority - Authority used in DARES tool?
 \item Aggregate - ?
 \item Status - ?
 \item TDIo - First diatom metric? Reference needed
 \item TDI3 - DARES river metric (previous)
 \item TDI4 - DARLEQ river metric (latest)
 \item LTDI1 - Previous Lake metric
 \item LTDI3 - DARLEQ lake metric (latest)
 \item DAM - Diatom Acid Metric
 \item Motile - Taxa classed as motile
 \item OrganicTolerant - Taxa classed as organic tolerant
 \item Planktic - Taxa classed as planktic
 \item Saline- Taxa classed as saline tolerant
 \item Notes - For example if typo or name change
}
}
\keyword{datasets}
